% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_strrugle_to_character.R
\name{class_strrugle_to_character}
\alias{class_strrugle_to_character}
\title{class_strrugle_to_character}
\usage{
class_strrugle_to_character(f_df, f_class)
}
\arguments{
\item{f_df}{f_df}

\item{f_class}{f_class}
}
\value{
return
}
\description{
\emph{class_strrugle_to_character}
}
\seealso{
\code{\link{class_strrugle_to_numeric}}
}
