% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_strrugle_to_numeric.R
\name{class_strrugle_to_numeric}
\alias{class_strrugle_to_numeric}
\title{}
\source{

}
\usage{
class_strrugle_to_numeric(f_df, f_class)
}
\arguments{
\item{f_df}{}

\item{f_class}{}
}
\value{

}
\description{

}
\section{Warning}{

}

\examples{

}
