% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_strrugle_to_character.R
\name{classStrrugleCharacter}
\alias{classStrrugleCharacter}
\title{classStrrugleCharacter}
\usage{
classStrrugleCharacter(fDf, fClass)
}
\arguments{
\item{fDf}{fDf}

\item{fClass}{fClass}
}
\value{
return
}
\description{
\emph{classStrrugleCharacter}
}
\seealso{
\code{\link{classStrrugleNumeric}}
}
