% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_ls.R
\name{concat_ls}
\alias{concat_ls}
\title{Concatenate List}
\usage{
concat_ls(f_data)
}
\arguments{
\item{f_data}{List of elements}
}
\value{
Concataned element of the same class as input
}
\description{
\emph{concat_ls} inputs a list of elements and concatenates them
}
