% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatLs.R
\name{concatLs}
\alias{concatLs}
\title{Concatenate List}
\usage{
concatLs(fData)
}
\arguments{
\item{fData}{List of elements}
}
\value{
Concataned element of the same class as input
}
\description{
\emph{concatLs} inputs a list of elements and concatenates them
}
