% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coorSeq.R
\name{coorSeq}
\alias{coorSeq}
\title{Coordinate Sequencer}
\usage{
coorSeq(fStartCoor, fEndCoor)
}
\arguments{
\item{fStartCoor}{Vector of start coordinates}

\item{fEndCoor}{Vector of end coordinates}
}
\value{
Vector sequence of concatenated integers in the same order as the coordinates
}
\description{
Generates a vector given a vector of start coordinates 'fStartCoor' and a vector of end coordinates 'fEndCoor'
}
\examples{
x <- sample(1:100, 10)
y <- x + sample(1:10, 10, replace = TRUE)
coorSeq(x, y)
}
