% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_tag.R
\name{dec_tag}
\alias{dec_tag}
\title{}
\source{

}
\usage{
dec_tag(f_seq, f_dec, f_numerator = 100)
}
\arguments{
\item{f_seq}{}

\item{f_dec}{}

\item{f_numerator}{}
}
\value{

}
\description{

}
\section{Warning}{

}

\examples{

}
