% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decListCompar.R
\name{decListCompar}
\alias{decListCompar}
\title{List Comparer}
\usage{
decListCompar(fVec1, fVec2, dSelector = 0)
}
\arguments{
\item{fVec1}{vector of decoded values}

\item{fVec2}{vector of decoded values}

\item{dSelector}{selector. \strong{Default = 0}}
}
\value{
vector of overlaping positions
}
\description{
Compare decoded lists \emph{fVec1} & \emph{fVec2}
}
