% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_clust.R
\name{h_clust}
\alias{h_clust}
\title{Hierarchical Cluster Tool}
\usage{
h_clust(f_dataset, f_pop)
}
\arguments{
\item{f_dataset}{matrix to values}

\item{f_pop}{population names}
}
\value{
return
}
\description{
\emph{h_clust} provides hierarchical clustering for group(s) of populations, i. e. it calculates through \emph{peak_diff} and constructs a matrix of the absolute differences among populations
It recives a data.frame of populations as \emph{f_dataset} where a vector of column names to be compare must be specified as \emph{f_pop}
}
\seealso{
\code{\link{locus_diff}}

\code{\link{peak_diff}}
}
