% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hClust.R
\name{hClust}
\alias{hClust}
\title{Hierarchical Cluster Tool}
\usage{
hClust(fDataset, fPop)
}
\arguments{
\item{fDataset}{matrix to values}

\item{fPop}{population names}
}
\value{
return
}
\description{
\emph{hClust} provides hierarchical clustering for group(s) of populations, i. e. it calculates through \emph{peakDiff} and constructs a matrix of the absolute differences among populations
It recives a data.frame of populations as \emph{fDataset} where a vector of column names to be compare must be specified as \emph{fPop}
}
\seealso{
\code{\link{locusDiff}}

\code{\link{peakDiff}}
}
