% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_diff.R
\name{locus_diff}
\alias{locus_diff}
\title{}
\source{

}
\usage{
locus_diff(f_x, f_y, f_data)
}
\arguments{
\item{f_x}{}

\item{f_y}{}

\item{f_data}{}
}
\value{

}
\description{

}
\section{Warning}{

}

\examples{

}
