% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_diff.R
\name{locusDiff}
\alias{locusDiff}
\title{locusDiff}
\usage{
locusDiff(fX, fY, fData)
}
\arguments{
\item{fX}{fX}

\item{fY}{fY}

\item{fData}{fData}
}
\value{
return
}
\description{
\emph{locusDiff}
}
\seealso{
\code{\link{peakDiff}}

\code{\link{hClust}}
}
