% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_junc_link.R
\name{mini_junc_link}
\alias{mini_junc_link}
\title{Mini Junction Linker}
\usage{
mini_junc_link(f_df, f_upp_col, f_down_col, f_pos_col, f_thres = 1)
}
\arguments{
\item{f_df}{data.frame}

\item{f_upp_col}{Column interpreted as upstream}

\item{f_down_col}{Column interpreted as downstream}

\item{f_pos_col}{Positions column}

\item{f_thres}{Threshold}
}
\value{
return
}
\description{
\emph{mini_junc_link}
}
