% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_junc_link.R
\name{miniJuncLink}
\alias{miniJuncLink}
\title{Mini Junction Linker}
\usage{
miniJuncLink(fDf, fUppCol, fDownCol, fPosCol, fThres = 1)
}
\arguments{
\item{fDf}{data.frame}

\item{fUppCol}{Column interpreted as upstream}

\item{fDownCol}{Column interpreted as downstream}

\item{fPosCol}{Positions column}

\item{fThres}{Threshold}
}
\value{
return
}
\description{
\emph{miniJuncLink}
}
