% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_diff.R
\name{peak_diff}
\alias{peak_diff}
\title{peak_diff}
\usage{
peak_diff(f_x, f_y, f_data)
}
\arguments{
\item{f_x}{f_x}

\item{f_y}{f_y}

\item{f_data}{f_data}
}
\value{
return
}
\description{
\emph{peak_diff}
}
\seealso{
\code{\link{locus_diff}}

\code{\link{h_clust}}

\code{\link{peak_iden}}

\code{\link{unique_coor}}

\code{\link{shared_coor}}
}
