% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_iden.R
\name{peak_iden}
\alias{peak_iden}
\title{Peak Identifier}
\usage{
peak_iden(f_seq, d_threshold = NULL)
}
\arguments{
\item{f_seq}{Numeric vector}

\item{d_threshold}{Threshold for peak identification. \strong{Default = 1}}
}
\value{
data.frame with following columns: \enumerate{
\item peak_no
\item lower_lim_ix
\item upper_lim_ix
\item peak_length_ix
}
}
\description{
Identifies peaks in a vector of number above a defined threshold \emph{'d_threshold'}
}
\examples{
x <- sample(0:5, 100, replace = TRUE, prob = c(5, rep(1, 5)))
peak_iden(x)
}
\seealso{
\code{\link{peak_iden_seq}}
}
