% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_iden_seq.R
\name{peak_iden_seq}
\alias{peak_iden_seq}
\title{Peak Identifier Plus}
\usage{
peak_iden_seq(f_seq, d_threshold = NULL, d_bin_size = NULL,
  d_bin_overlaps = NULL)
}
\arguments{
\item{f_seq}{Numeric vector}

\item{d_threshold}{Threshold for peak identification. \strong{Default = 1}}

\item{d_bin_size}{Defines bin size assuming \emph{f_seq} is a vector of sliding windows. \strong{Default = 500}}

\item{d_bin_overlaps}{Defines bin overlaps assuming \emph{f_seq} is a vector of sliding windows. \strong{Default = 10}}
}
\value{
data.frame with following columns: \enumerate{
\item peak_no
\item seq_max
\item seq_min
\item seq_mean
\item seq_sum
\item lower_lim_ix
\item upper_lim_ix
\item peak_length_ix
\item peak_length_nt
}
}
\description{
Identifies peaks in a vector of number above a defined threshold \emph{'d_threshold'}.
Additionally, calculates \emph{'min'}, \emph{'mean'}, \emph{'sum'} and coordinates in indixes (ix) and nucleotides (nt) given \emph{'d_bin_size'} and \emph{'d_bin_overlaps'} assuming it is dealing with overlaping windows.
}
\examples{
x <- sample(0:5, 100, replace = TRUE, prob = c(5, rep(1, 5)))
peak_iden_seq(
x,
d_threshold = 1,
d_bin_size = 10,
d_bin_overlaps = 1
)
}
\seealso{
\code{\link{peak_iden}}
}
