% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_tag.R
\name{peak_tag}
\alias{peak_tag}
\title{}
\source{

}
\usage{
peak_tag(f_df_pos, f_seq_dist)
}
\arguments{
\item{f_df_pos}{}

\item{f_seq_dist}{}
}
\value{

}
\description{

}
\section{Warning}{

}

\examples{

}
