% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakDiff.R
\name{peakDiff}
\alias{peakDiff}
\title{peakDiff}
\usage{
peakDiff(fX, fY, fData)
}
\arguments{
\item{fX}{fX}

\item{fY}{fY}

\item{fData}{fData}
}
\value{
return
}
\description{
\emph{peakDiff}
}
\seealso{
\code{\link{locusDiff}}

\code{\link{hClust}}

\code{\link{peakIden}}

\code{\link{uniqueCoor}}

\code{\link{sharedCoor}}
}
