% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_iden.R
\name{peakIden}
\alias{peakIden}
\title{Peak Identifier}
\usage{
peakIden(fSeq, dThreshold = NULL)
}
\arguments{
\item{fSeq}{Numeric vector}

\item{dThreshold}{Threshold for peak identification. \strong{Default = 1}}
}
\value{
data.frame with following columns: \enumerate{
\item peakNo
\item lowerLimIx
\item upperLimIx
\item peakLengthIx
}
}
\description{
Identifies peaks in a vector of number above a defined threshold \emph{'dThreshold'}
}
\examples{
x <- sample(0:5, 100, replace = TRUE, prob = c(5, rep(1, 5)))
peakIden(x)
}
\seealso{
\code{\link{peakIdenSeq}}
}
