% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakIdenSeq.R
\name{peakIdenSeq}
\alias{peakIdenSeq}
\title{Peak Identifier Plus}
\usage{
peakIdenSeq(fSeq, dThreshold = NULL, dBinSize = NULL, dBinOverlaps = NULL)
}
\arguments{
\item{fSeq}{Numeric vector}

\item{dThreshold}{Threshold for peak identification. \strong{Default = 1}}

\item{dBinSize}{Defines bin size assuming \emph{fSeq} is a vector of sliding windows. \strong{Default = 500}}

\item{dBinOverlaps}{Defines bin overlaps assuming \emph{fSeq} is a vector of sliding windows. \strong{Default = 10}}
}
\value{
data.frame with following columns: \enumerate{
\item peakNo
\item seqMax
\item seqMin
\item seqMean
\item seqSum
\item lowerLimIx
\item upperLimIx
\item peakLengthIx
\item peakLengthNt
}
}
\description{
Identifies peaks in a vector of number above a defined threshold \emph{'dThreshold'}.
Additionally, calculates \emph{'min'}, \emph{'mean'}, \emph{'sum'} and coordinates in indixes (ix) and nucleotides (nt) given \emph{'dBinSize'} and \emph{'dBinOverlaps'} assuming it is dealing with overlaping windows.
}
\examples{
x <- sample(0:5, 100, replace = TRUE, prob = c(5, rep(1, 5)))
peakIdenSeq(
x,
dThreshold = 1,
dBinSize = 10,
dBinOverlaps = 1
)
}
\seealso{
\code{\link{peakIden}}
}
