% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_assembler.R
\name{range_assembler}
\alias{range_assembler}
\title{Range Assembler}
\usage{
range_assembler(f_data, b_genom = TRUE)
}
\arguments{
\item{f_data}{data.frame to be transform. Must be formated as: ranges start, ranges end and seqnames}

\item{b_genom}{boolean to set if chromosome values are set}
}
\value{
GenomicRanges object
}
\description{
Wrapper to create GenomicRanges objects in one line
}
\seealso{
\code{\link{unique_coor}}

\code{\link{shared_coor}}
}
