% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_assembler.R
\name{rangeAssembler}
\alias{rangeAssembler}
\title{Range Assembler}
\usage{
rangeAssembler(fData, bGenom = TRUE)
}
\arguments{
\item{fData}{data.frame to be transform. Must be formated as: ranges start, ranges end and seqnames}

\item{bGenom}{boolean to set if chromosome values are set}
}
\value{
GenomicRanges object
}
\description{
Wrapper to create GenomicRanges objects in one line
}
\seealso{
\code{\link{uniqueCoor}}

\code{\link{sharedCoor}}
}
