% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounder.R
\name{rounder}
\alias{rounder}
\title{Rounder}
\source{

}
\usage{
rounder(f_seq, f_round_factor)
}
\arguments{
\item{f_seq}{Vector of numbers}

\item{f_round_factor}{Rounding factor}
}
\value{
Returns a vector of numbers
}
\description{
Rounds numbers 'f_seq' to ceiling given a factor 'f_round_factor'
}
\section{Warning}{

}

\examples{
rounder(27, 10)
rounder(341:710, 20)
}
