% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slid_win.R
\name{slid_win}
\alias{slid_win}
\title{Window Slider}
\usage{
slid_win(f_seq, d_bin_size = 500)
}
\arguments{
\item{f_seq}{Numerical vector}

\item{d_bin_size}{Bin size. \strong{Default = 500}}
}
\value{
Numerical vector
}
\description{
Takes in a vector of numbers \emph{'f_seq'} and outputs its positions assuming \emph{'d_bin_size'}
}
\examples{
slid_win(37, 5)
slid_win(1:100, 15)
}
