% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slid_win_ov.R
\name{slid_win_ov}
\alias{slid_win_ov}
\title{Window Slider with Overlaps}
\usage{
slid_win_ov(f_seq, d_bin_size = 500, d_bin_overlap = 10)
}
\arguments{
\item{f_seq}{Numerical vector}

\item{d_bin_size}{Bin size. \strong{Default = 500}}

\item{d_bin_overlap}{Bin overlap. \strong{Default = 10}}
}
\value{
Numerical vector
}
\description{
Takes in numerical vector as \emph{'f_seq'} and outputs their positions assuming \emph{'d_bin_size'} and \emph{'d_bin_overlap'}
}
\section{Warning}{

\emph{'slid_win_ov'} is only useful for single numbers, otherwise use \emph{'slid_win_tov'}
\emph{'slid_win_ov'} uses \emph{'slid_win'} to determine the window sliding
}

\examples{
slid_win_ov(37, 15, 5)
}
\seealso{
\code{\link{slid_win}}
}
