% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slid_win_tov.R
\name{slid_win_tov}
\alias{slid_win_tov}
\title{Window Slider with Table Overlaps}
\usage{
slid_win_tov(f_seq, d_bin_size = 500, d_bin_overlap = 10)
}
\arguments{
\item{f_seq}{Numerical vector}

\item{d_bin_size}{Bin size. \strong{Default = 500}}

\item{d_bin_overlap}{Bin overlap. \strong{Default = 10}}
}
\value{
Table with overlap indexes
}
\description{
Takes in a numerical vector \emph{'f_seqs'} and outputs their positions assuming \emph{'d_bin_size'} and \emph{'d_bin_overlap'}
}
\section{Warning}{

\emph{'slid_win_ov'} uses \emph{'slid_win'} to determine the window sliding
}

\examples{
slid_win_tov(1:100, 25, 5)
}
\seealso{
\code{\link{slid_win}}
}
