% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slid_win.R
\name{slidWin}
\alias{slidWin}
\title{Window Slider}
\usage{
slidWin(fSeq, dBinSize = 500)
}
\arguments{
\item{fSeq}{Numerical vector}

\item{dBinSize}{Bin size. \strong{Default = 500}}
}
\value{
Numerical vector
}
\description{
Takes in a vector of numbers \emph{'fSeq'} and outputs its positions assuming \emph{'dBinSize'}
}
\examples{
slidWin(37, 5)
slidWin(1:100, 15)
}
