% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slid_win_tov.R
\name{slidWinTov}
\alias{slidWinTov}
\title{Window Slider with Table Overlaps}
\usage{
slidWinTov(fSeq, dBinSize = 500, dBinOverlap = 10)
}
\arguments{
\item{fSeq}{Numerical vector}

\item{dBinSize}{Bin size. \strong{Default = 500}}

\item{dBinOverlap}{Bin overlap. \strong{Default = 10}}
}
\value{
Table with overlap indexes
}
\description{
Takes in a numerical vector \emph{'fSeqs'} and outputs their positions assuming \emph{'dBinSize'} and \emph{'dBinOverlap'}
}
\section{Warning}{

\emph{'slidWinOv'} uses \emph{'slidWin'} to determine the window sliding
}

\examples{
slidWinTov(1:100, 25, 5)
}
\seealso{
\code{\link{slidWin}}
}
