% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_collector.R
\name{tb_collector}
\alias{tb_collector}
\title{}
\source{

}
\usage{
tb_collector(f_data, f_ix, f_exp_val_range, f_val = NULL)
}
\arguments{
\item{f_data}{}
}
\value{

}
\description{
Collects vector data into a table of standard format defined by expected values
}
\section{Warning}{

}

\examples{

}
