% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_collector.R
\name{tbCollector}
\alias{tbCollector}
\title{tbCollector}
\usage{
tbCollector(fData, fIx, fExpValRange, fVal = NULL)
}
\arguments{
\item{fData}{fData}

\item{fIx}{fIx}

\item{fExpValRange}{expected table range}

\item{fVal}{fVal}
}
\value{
return
}
\description{
\emph{tbCollector} collects vector data into a table of standard format defined by expected values
}
\seealso{
\code{\link{concatLs}}
}
