% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueCoor.R
\name{uniqueCoor}
\alias{uniqueCoor}
\title{Unique Coordinate}
\usage{
uniqueCoor(fQuery, fSubj, query, subj, dGenomic = FALSE)
}
\arguments{
\item{fQuery}{data.frame interpreted as query with specified format}

\item{fSubj}{data.frame interpreted as subject with specified format}

\item{query}{Names to assign to query on output}

\item{subj}{Names to assign to subject on output}

\item{dGenomic}{boolean as to interpret third column as chromosome}
}
\value{
List of query and subject locations.
}
\description{
\emph{uniqueCoor} takes two data.frames with start (first column) and end (second column) coordinates, and finds unique positions.
\emph{dGenomic} option ( default = \emph{FALSE} ) if non-genomic coordinates, otherwise chromosome information must be entered in column 3.
requires \strong{GenomicRanges} (built-in).
}
\seealso{
\code{\link{sharedCoor}}
}
