% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classStrrugleNumeric.R
\name{classStrrugleNumeric}
\alias{classStrrugleNumeric}
\title{classStrrugleNumeric}
\usage{
classStrrugleNumeric(fDf, fClass)
}
\arguments{
\item{fDf}{fDf}

\item{fClass}{fClass}
}
\value{
return
}
\description{
\emph{classStrrugleNumeric}
}
\seealso{
\code{\link{classStrrugleCharacter}}
}
