% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec_tag.R
\name{dec_tag}
\alias{dec_tag}
\title{Tag Decoder}
\usage{
dec_tag(f_seq, f_dec, d_numerator = 100)
}
\arguments{
\item{f_seq}{Numerical vector}

\item{f_dec}{Position to decode}

\item{d_numerator}{Digit position}
}
\value{
Numerical vector of decoded values
}
\description{
\emph{dec_tag} decodes a numeric vector given a specified position \emph{f_dec} by digits as \emph{d_numerator}
}
\examples{

}
