% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_diff.R
\name{locus_diff}
\alias{locus_diff}
\title{locus_diff}
\usage{
locus_diff(f_x, f_y, f_data)
}
\arguments{
\item{f_x}{f_x}

\item{f_y}{f_y}

\item{f_data}{f_data}
}
\value{
return
}
\description{
\emph{locus_diff}
}
\seealso{
\code{\link{peak_diff}}

\code{\link{h_clust}}
}
