% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_junc_link.R
\name{mini_junc_link}
\alias{mini_junc_link}
\title{}
\source{

}
\usage{
mini_junc_link(f_df, f_upp_col, f_down_col, f_pos_col, f_thres = 1,
  f_chrom_col = NULL)
}
\arguments{
\item{f_df}{}

\item{f_upp_col}{}

\item{f_down_col}{}

\item{f_pos_col}{}

\item{f_thres}{}

\item{f_chrom_col}{}
}
\value{

}
\description{

}
\section{Warning}{

}

\examples{

}
