% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_diff.R
\name{peak_diff}
\alias{peak_diff}
\title{}
\source{

}
\usage{
peak_diff(f_x, f_y, f_data)
}
\arguments{
\item{f_x}{}

\item{f_y}{}

\item{f_data}{}
}
\value{

}
\description{

}
\section{Warning}{

}

\examples{

}
