% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_iden.R
\name{peak_iden}
\alias{peak_iden}
\title{Peak Identifier}
\source{

}
\usage{
peak_iden(f_seq, f_threshold = NULL)
}
\arguments{
\item{f_seq}{Vector of numbers}

\item{f_threshold}{Threshold for peak identification. Default = 1}
}
\value{
Returns a data.frame with following columns: "peak_no", "lower_lim_ix", "upper_lim_ix", "peak_length_ix"
}
\description{
Identifies peaks in a vector of number above a defined threshold 'f_threshold'
}
\section{Warning}{

}

\examples{
x <- sample(0:5, 100, replace=T, prob = c(5, rep(1, 5)))
peak_iden_seq(x)
}
