% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_iden_seq.R
\name{peak_iden_seq}
\alias{peak_iden_seq}
\title{Peak Sequence Identifier}
\source{

}
\usage{
peak_iden_seq(f_seq, f_threshold = NULL, f_bin_size = NULL,
  f_bin_overlaps = NULL)
}
\arguments{
\item{f_seq}{Vector of numbers}

\item{f_threshold}{Threshold for peak identification. Default = 1}

\item{f_bin_size}{Defines bin size assuming f_seq is a vector of sliding windows. Default = 'bin_size'}

\item{f_bin_overlaps}{Defines bin overlaps assuming f_seq is a vector of sliding windows. Default = 'bin_overlaps'}
}
\value{
Returns a data.frame with following columns: "peak_no", "seq_max", "seq_min", "seq_mean", "seq_sum", "lower_lim_ix", "upper_lim_ix", "peak_length_ix", "peak_length_nt"
}
\description{
Identifies peaks in a vector of number above a defined threshold 'f_threshold'. Additionally, calculates min, mean, sum and coordinates in indixes (ix) and nucleotides (nt) given 'f_bin_size' and 'f_bin_overlaps' assuming it is dealing with overlaping windows.
}
\section{Warning}{

}

\examples{
x <- sample(0:5, 100, replace=T, prob = c(5, rep(1, 5)))
peak_iden_seq(
x,
f_threshold = 1,
f_bin_size = 10,
f_bin_overlaps = 1
)
}
