% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_tag.R
\name{peak_tag}
\alias{peak_tag}
\title{peak_tag}
\usage{
peak_tag(f_df_pos, f_seq_dist)
}
\arguments{
\item{f_df_pos}{f_df_pos}

\item{f_seq_dist}{f_seq_dist}
}
\value{
return
}
\description{
\emph{peak_tag}
}
