% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounder.R
\name{rounder}
\alias{rounder}
\title{Rounder}
\usage{
rounder(f_seq, f_round_factor)
}
\arguments{
\item{f_seq}{Numerical vector}

\item{f_round_factor}{Rounding factor}
}
\value{
Numerical vector
}
\description{
Rounds numbers \emph{'f_seq'} to ceiling given a factor \emph{'f_round_factor'}
}
\examples{
rounder(27, 10)
rounder(341:710, 20)
}
