% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounder.R
\name{rounder}
\alias{rounder}
\title{Rounder}
\usage{
rounder(fSeq, fRoundFactor)
}
\arguments{
\item{fSeq}{Numerical vector}

\item{fRoundFactor}{Rounding factor}
}
\value{
Numerical vector
}
\description{
Rounds numbers \emph{'fSeq'} to ceiling given a factor \emph{'fRoundFactor'}
}
\examples{
rounder(27, 10)
rounder(341:710, 20)
}
