% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_coor.R
\name{shared_coor}
\alias{shared_coor}
\title{}
\source{

}
\usage{
shared_coor(f_query, f_subj, query, subj, f_genomic = F)
}
\arguments{
\item{f_query}{}

\item{f_subj}{}

\item{query}{}

\item{subj}{}

\item{f_genomic}{}
}
\value{

}
\description{

}
\section{Warning}{

}

\examples{

}
