% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slid_win.R
\name{slid_win}
\alias{slid_win}
\title{Window Slider}
\source{

}
\usage{
slid_win(f_seq, f_bin_size = bin_size)
}
\arguments{
\item{f_seq}{Vector of numbers}

\item{f_bin_size}{Bin size}
}
\value{
Returns a vector of numbers
}
\description{
Takes in a vector of numbers 'f_seq' and outputs its positions assuming 'f_bin_size'
}
\section{Warning}{

}

\examples{
slid_win(37, 5)
slid_win(1:100, 15)
}
