% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slid_win_ov.R
\name{slid_win_ov}
\alias{slid_win_ov}
\title{Window Slider with Overlaps}
\source{

}
\usage{
slid_win_ov(f_seq, f_bin_size = bin_size, f_bin_overlap = bin_overlaps)
}
\arguments{
\item{f_seq}{Vector of numbers}

\item{f_bin_size}{Bin size}

\item{f_bin_overlap}{Bin overlap}
}
\value{
Returns a vector of numbers
}
\description{
Takes in a vector of numbers 'f_seq' and outputs their positions assuming 'f_bin_size' and 'f_bin_overlap'
}
\section{Warning}{

'slid_win_ov' is only useful for single numbers, otherwise use 'slid_win_tov'
'slid_win_ov' uses 'slid_win' to determine the window sliding
}

\section{See also}{

'slid_win'
}

\examples{
slid_win_ov(37, 15, 5)
}
