% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slid_win_tov.R
\name{slid_win_tov}
\alias{slid_win_tov}
\title{Window Slider with Overlaps}
\source{

}
\usage{
slid_win_tov(f_seqs, f_bin_size = bin_size,
  f_bin_overlap = bin_overlaps)
}
\arguments{
\item{f_bin_size}{Bin size}

\item{f_bin_overlap}{Bin overlap}

\item{f_seq}{Vector of numbers}
}
\value{
Returns a table with overlap indexes
}
\description{
Takes in a vector of numbers 'f_seqs' and outputs their positions assuming 'f_bin_size' and 'f_bin_overlap'
}
\section{Warning}{

'slid_win_ov' uses 'slid_win' to determine the window sliding
}

\section{See also}{

'slid_win'
}

\examples{
slid_win_tov(1:100, 25, 5)
}
